+++ lib/entity/friend_info_entity.dart
@@ -25,33 +25,30 @@ class FriendInfoEntity {
   });
 
   FriendInfoEntity.fromJson(data) {
-    Map<String, dynamic> json =
-        data is Map ? data.cast<String, dynamic>() : jsonDecode(data);
+    Map<String, dynamic> json = data is Map ? data.cast<String, dynamic>() : jsonDecode(data);
     if (json['userID'] != null) userID = json['userID'];
     if (json['friendRemark'] != null) friendRemark = json['friendRemark'];
-    if (json['friendGroups'] != null)
-      friendGroups = json['friendGroups']?.cast<String>();
-    if (json['friendCustomInfo'] != null)
-      friendCustomInfo = json['friendCustomInfo']?.cast<String, String>();
-    if (json['userProfile'] != null)
-      userProfile = UserEntity.fromJson(json['userProfile']);
+    if (json['friendGroups'] != null) friendGroups = json['friendGroups']?.cast<String>();
+    if (json['friendCustomInfo'] != null) {
+      Map<String, String> _customInfo = {};
+      (json['friendCustomInfo'] as Map).cast<String, String>().forEach((key, value) {
+        _customInfo[key] = (value ?? "").replaceAll("Tag_Profile_Custom_", "");
+      });
+      friendCustomInfo = _customInfo;
+    }
+    if (json['userProfile'] != null) userProfile = UserEntity.fromJson(json['userProfile']);
   }
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
     data['userID'] = this.userID;
     if (this.friendRemark != null) data['friendRemark'] = this.friendRemark;
-    if (this.friendCustomInfo != null)
-      data['friendCustomInfo'] = this.friendCustomInfo;
+    if (this.friendCustomInfo != null) data['friendCustomInfo'] = this.friendCustomInfo;
     return data;
   }
 
   @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is FriendInfoEntity &&
-          runtimeType == other.runtimeType &&
-          userID == other.userID;
+  bool operator ==(Object other) => identical(this, other) || other is FriendInfoEntity && runtimeType == other.runtimeType && userID == other.userID;
 
   @override
   int get hashCode => userID.hashCode;